STRUCTURE::CHANGE:Cleaning the project structure
