+++ Rasp/Test/testLora.py
@@ -7,7 +7,7 @@ destNet = str(input('Enter the target network number : '))
 
 if True:
     print("Initializing LoRa module")
-    proc = subprocess.Popen(["../Rasp/Init"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+    proc = subprocess.Popen(["../../Rasp/Init"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
     print(proc)  # initializing lora module
 
     stdout, stderr = proc.communicate(timeout=15)
@@ -17,7 +17,7 @@ if True:
 
 if False:
     print("Transmitting a LED_ON order")
-    proc = subprocess.Popen(["../Rasp/Transmit", "T", destNet, myNet, "1", "1", "1"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+    proc = subprocess.Popen(["../../Rasp/Transmit", "T", destNet, myNet, "1", "1", "1"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
     print(proc) # turn on the led
 
     stdout, stderr = proc.communicate(timeout=15)
@@ -27,7 +27,7 @@ sleep(5)
 
 if True:
     print("Waiting for a LoRa packet")
-    proc = subprocess.Popen(["../Rasp/Receive", myNet, "20"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+    proc = subprocess.Popen(["../../Rasp/Receive", myNet, "20"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
     print(proc)  # Waiting for zigbee ack
 
     stdout, stderr = proc.communicate(timeout=60)
@@ -39,7 +39,7 @@ sleep(5)
 
 if False:
     print("Transmitting a LED_OFF order")
-    proc = subprocess.Popen(["../Rasp/Transmit", "T", destNet, myNet, "1", "1", "0"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+    proc = subprocess.Popen(["../../Rasp/Transmit", "T", destNet, myNet, "1", "1", "0"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
     print(proc)  # shut off the led
 
     stdout, stderr = proc.communicate(timeout=15)
@@ -49,7 +49,7 @@ sleep(5)
 
 if True:
     print("Waiting for a LoRa packet")
-    proc = subprocess.Popen(["../Rasp/Receive", 1, "20"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+    proc = subprocess.Popen(["../../Rasp/Receive", 1, "20"], shell=False, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
     print(proc)  # waiting for zigbee ack
 
     stdout, stderr = proc.communicate(timeout=60)
