--- Zolertia/script
@@ -1,10 +1,0 @@
-git clone https://github.com/contiki-ng/contiki-ng.git
-sudo apt update
-cd contiki-ng
-git submodule update --init
-sudo apt install python3-pip
-sudo pip3 install pyserial
-sudo pip3 install intelhex
-sudo pip3 install python-magic
-sudo pip3 install grovepi
-sudo pip3 install pytz
