+++ script
@@ -1,0 +1,32 @@
+echo "bind 'set completion-ignore-case on'" >> ~/.bashrc
+sudo apt update; sudo apt full-upgrade -y; sudo apt autoremove -y; sudo apt purge; sudo apt clean
+curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+sudo apt install -y libgirepository1.0-dev
+sudo apt install -y libdbus-1-dev
+sudo apt install -y libglib2.0-dev
+sudo apt install -y libcairo2-dev
+sudo apt install -y libssl-dev
+pip3 install -U pip
+pip3 list -o | cut -f1 -d' ' | tr " " "\n" | awk '{if(NR>=3)print}' | cut -d' ' -f1 | xargs -n1 pip3 install -U > /dev/tty1
+
+
+
+git clone https://github.com/contiki-ng/contiki-ng.git
+cd contiki-ng
+git submodule update --init
+sudo apt install -y python3-pip
+sudo apt install -y gcc-arm-none-eabi
+sudo pip3 install pyserial
+sudo pip3 install intelhex
+sudo pip3 install python-magic
+sudo pip3 install grovepi
+sudo pip3 install pytz
+sudo chmod 666 /dev/ttyUSB0
+# cp ~/Zigbee-LoRa-Interconnection/Zolertia/zolertia_Z1.c ~/contiki-ng/examples/rpl-udp/
+cp ~/zolertia_Z1.c ~/contiki-ng/examples/rpl-udp/
+cd ~/contiki-ng/examples/rpl-udp/
+sudo make BOARD=firefly TARGET=zoul zolertia_ZN.upload login
+# ../../tools/cc2538-bsl/cc2538-bsl.py -e -w -v -a 0x00202000 \
+# ../../tools/serial-io/serialdump -b115200 /dev/ttyUSB0
+# {"ID":1,"T":1,"O":1,"NETD":2,"NETS":1}
+# python raspb_Z1.py
